Made Modifications but Came with Major Bugs
Okay it still seems to work on the website view, but here are the
problems I want to note from the mobile perspective.

1) After creating a new course on the home page when you click on the
course, the font size decreases and he box itself shrinks.
2) When creating a new assignment inside a course, the format is a
little funny.... For example, it looks like
Assignments(10%)0% instead of
Assignments(10%)      0% because I'm assuming that we want some space
between the name of the assignment and the % in the class.
3) When clicking on an assignment, the same thing happens where the size
of the text shrinks, and I can't seem to figure out why... I took a look
at the hover aspects in the CSS and I think the problem is due to a
conflict between that code and the new code I added.

4) When you have two or more assignments and click the home page, there
is a momentary big blank space on the screen.

5) After doing step 4, when you revisit the page, the blank space
actually stays in place.

It's bad -- it's real bad....