So Done With this BullShit
Okay so most of the changes are fixed and I made it look as close to
possible to before we started packaging. However, could you please fix
the "item delete" option -- for iphone 4 and iphone 5, the "no" goes
onto the second line and that's ugly. Also, I couldn't figure out how to
fix the header because am too noob, so pls help.